migrate com.google.fonts/check/032 to googlefonts spec
(issue #1956 and also issue #1945)
