+++ Lib/fontbakery/specifications/googlefonts.py
@@ -847,6 +847,27 @@ def com_google_fonts_check_030(ttFont, familyname):
         yield PASS, "Font has a valid license URL in NAME table."
 
 
+@check(
+  id = 'com.google.fonts/check/032'
+)
+def com_google_fonts_check_032(ttFont):
+  """Description strings in the name table must not exceed 100 characters."""
+  from fontbakery.constants import NAMEID_DESCRIPTION
+  failed = False
+  for name in ttFont['name'].names:
+    if (name.nameID == NAMEID_DESCRIPTION and
+        len(name.string.decode(name.getEncoding())) > 100):
+      failed = True
+      break
+
+  if failed:
+    yield FAIL, ("Namerecords with ID={} (NAMEID_DESCRIPTION)"
+                 " are longer than 100 characters"
+                 " and should be removed.").format(NAMEID_DESCRIPTION)
+  else:
+    yield PASS, "Description name records do not exceed 100 characters."
+
+
 @condition
 def ttfautohint_stats(font):
   import re
